WebControllers update and small html fix
I added the WebControllers for the last two classes and the repository. I also noticed in my testing that when we were in the view html page it wouldn't call back to the input page. I add ../ to it to fix this. There is still an issue with the program, when the user tries to input a session it asks for username. However, since we are using OneToOne it now checks for the id instead of the username. We might be able to change it to use username as the primary key or something to get it to work.