Login Implementation
I added the login using the @SessionAttribute tag. I have updated the WebControllers for the Exercise and Users class to only show the data of the person currently logged in. I made a very basic login.html page that just tells the user to log in and takes username and password. I think we should make a login page the index.html like we discussed and make a link allowing them to create an account (Users object) if they don't have one. We can keep the view all users so they can see and edit their account. I didn't want to change too much and mess Fernando up on the html so the only changes I made were a few things needed to link differently and I added a link to the login page on the index.html page. I also added the new variables I put in our classes to the html pages.