Various Fixes and Changes (Details in description)
Linked the CSS and logo png to HTML pages. and fixed a bug with the Meals WebControllers linking to the input HTML pages. Removed user_id so the user can't enter it or see it (everything that is needed for is done in the WebControllers). Changed Home to Logout and added a Profile link to nav bar.  Swapped dayJoined in Users back to LocalDate and made it so the WebControllers just keeps track of when the user joined (it should keep the date they joined even if they edit their profile) (we may have to drop and rebuild the Users table because of it).  Implemented a home page/main menu HTML page that tells the user basic facts about FitTrack. Changed Index HTML to be a page linking to the Login and Create Users page for HTML and WebController linking purposes. (I copied our previous index.html that was the login page and changed it to login in templates (being in static was producing an error with the WebControllers I think).  Changed where various links lead to so that the user can't bypass login.